made some changes

+++ blueprints/expense/routes.py
@@ -7,9 +7,9 @@ from models.expense import Expense
 from flask_jwt_extended import create_access_token, create_refresh_token, get_jwt_identity, jwt_required
 from blueprints.expense.utils import get_expense_by_date
 
-@expense_bp.route('', methods={'POST'})
+@expense_bp.route('', methods=['POST'])
 @jwt_required()
-def add_expense(user_id):
+def add_expense:
     #adds new expense
     # gets the user id of the logged in user
     user_id = get_jwt_identity()
@@ -17,11 +17,11 @@ def add_expense(user_id):
     # gets the expesne data from the request
     data = request.get_json()
     if not data or not 'amount' in data or not  'category' in data:
-        return jsonify({'error': 'Mising required expenses fields'}), 400
+        return jsonify({'error': 'Missing required expenses fields'}), 400
     
     amount = data['amount']
     category = data['category']
-    description = data['description']
+    description = data.get('description')
     
     try:
         newExpense = Expense(user_id=user_id, amount=amount, category=category, description=description)
@@ -36,19 +36,18 @@ def add_expense(user_id):
 @jwt_required()
 def get_expenses():
     user_id = get_jwt_identity()
-    data = request.get_json()
-    if not data or not 'filter_type' in data:
+    data = request.args.get('filter_type')
+    if  not 'filter_type':
         return jsonify({'error': 'Mising required expenses filter fields'}), 400
     
-    filter = data['filter_type']
-    if filter is 'custom':
-        if not 'start_date' in data or not 'end_date' in data:
-            return jsonify({'error': 'Mising required expenses filter fields (start or end date)'}), 400
-        else:
-            start_date = data['start_date']
-            end_date = data['end_date']
+    if filter_type == 'custom':
+        start_date = request.args.get("start_date")
+        end_date = request.args.get("end_date")
+
+        if not start_date or not end_date:
+            return jsonify({"error": "Missing start_date or end_date for custom filter"}), 400
             try:
-                resonse = get_expense_by_date(user_id, filter, start_date, end_date)
+                response = get_expense_by_date(user_id, filter, start_date, end_date)
                 return jsonify(response), 200
             except Exception as err:
                 return jsonify({'error': str(err)}), 400
@@ -65,8 +64,8 @@ def delete_expense(expense_id):
     """Delete an expense"""
     user_id = get_jwt_identity()
     expense = storage.get_expense(expense_id)
-    if not expense:
-        return jsonify({"error": "Expense not found"}), 404
+    if not expense or expense.user_id != user_id:
+        return jsonify({"error": "Unauthorised or Expense not found"}), 404
     storage.delete(expense)
     storage.save()
     return jsonify({"message": f"Expense {expense.id} deleted  successfully"}), 200
@@ -77,11 +76,11 @@ def update_expense(expense_id):
     """Update an expense"""
     user_id = get_jwt_identity()
     expense = storage.get_expense(expense_id)
-    if not expense:
-        return jsonify({"error": "Expense not found"}), 404
+    if not expense or expense.user_id != user_id:
+        return jsonify({"error": "Unauthorised or Expense not found"}), 404
 
     data = request.get_json()
-    if "amount" in data:
+    if "amount" in data and amount > 0:
         expense.amount = data["amount"]
     if "category" in data:
         expense.category = data["category"]
@@ -89,4 +88,4 @@ def update_expense(expense_id):
         expense.description = data["description"]
 
     storage.save()
-    return jsonify({"message": "Expense updated successfully"}), 200
+    return jsonify({"message": "Expense updated successfully", "expense": expense.to_dict()}), 200
